+++ scripts/dice-stats-db.js
@@ -5,24 +5,24 @@ class DB_INTERACTION
      */
     static createDB()
     {
-        for(let user in game.users)
-        {
-            if(!hasProperty(user, 'data.flags.'+MODULE_ID+'.'+ROLLDATAFLAG))
-            {
-                user.setFlag(MODULE_ID,ROLLDATAFLAG,{});
-            }
-        }
+        //Not needed
     }
 
     /**
-     * Method called to save user info to the db
+     * Method called to save user info to the db, Can only save your own data!
      * @param {PLAYER} playerInfo 
      */
     static saveUserData(playerInfo)
     {
+        let userid = playerInfo.USERID;
         let objCpy = Object.assign({},playerInfo);
-        game.user.setFlag(MODULE_ID,'player_roll_data', objCpy);
+        if(userid)
+        {
+            game.users.get(userid)?.setFlag(MODULE_ID,'player_roll_data', objCpy);
+        }
     }
+
+
 
     /**
      * Method used to load all players info from the db
@@ -32,13 +32,12 @@ class DB_INTERACTION
         let user = game.users.get(userId);
         if(user)
         {
-            if(hasProperty(user, 'data.flags.'+MODULE_ID+'.player_roll_data'))
+            if(user.getFlag(MODULE_ID,'player_roll_data'))
             {
                 return user.getFlag(MODULE_ID,'player_roll_data');
             }
-            return null;
         }
-        
+        return null;
     }
 
     //Method used to convert DB info to A player OBJ
@@ -103,11 +102,14 @@ class DB_INTERACTION
      */
     static clearDB()
     {
-        for(let user in game.users)
+        for(let aUser of game.users)
         {
-            if(hasProperty(user, 'data.flags.'+MODULE_ID+'.player_roll_data'))
+            if(aUser)
             {
-                user.unsetFlag(MODULE_ID,'player_roll_data');
+                if(aUser.getFlag(MODULE_ID,'player_roll_data'))
+                {
+                    aUser.unsetFlag(MODULE_ID,'player_roll_data');
+                }
             }
         }
     } 
@@ -117,9 +119,12 @@ class DB_INTERACTION
      */
     static clearPlayer(user)
     {
-        if(hasProperty(user, 'data.flags.'+MODULE_ID+'.player_roll_data'))
+        if(user)
         {
-            user.unsetFlag(MODULE_ID,'player_roll_data');
+            if(user.getFlag(MODULE_ID,'player_roll_data'))
+            {
+                user.unsetFlag(MODULE_ID,'player_roll_data');
+            }
         }
     }
 }