+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG  
 
+## [ 1.10.0 ] 5/8/2023
+### Added
+- Updated the stats to display the decimal value of the average rather then the integer value
+- Fix clear All Data from DB buttons not working correctly
+### Updated
+- Changed Default of Auto DB to be **ON**
+- Changed new Buttons to be **ON** by default
+
 ## [ 1.9.1 ]  4/28/2023
 ### Fixed
 - Bugfix for gm visibility setting